python/requirements.txt
aiofiles~=23.1.0
httpx~=0.25.0
openai~=0.28.0
aiogram~=3.1.1
python-dotenv~=1.0.0